extract account into separate package
